socket io api use ssize_t

+++ inc/sysapi/socket.h
@@ -155,9 +155,9 @@ __declspec_dll FD_t socketTcpListen2(int family, const char* ip, unsigned short 
 __declspec_dll FD_t socketTcpAccept(FD_t listenfd, int msec, struct sockaddr* from, socklen_t* p_slen);
 __declspec_dll BOOL socketPair(int type, FD_t sockfd[2]);
 __declspec_dll int socketRecvFrom(FD_t sockfd, void* buf, unsigned int buflen, int flags, struct sockaddr* from, socklen_t* p_slen);
-__declspec_dll int socketReadv(FD_t sockfd, Iobuf_t iov[], unsigned int iovcnt, int flags, struct sockaddr* from, socklen_t* p_slen);
-__declspec_dll int socketWritev(FD_t sockfd, const Iobuf_t iov[], unsigned int iovcnt, int flags, const struct sockaddr* to, socklen_t tolen);
-__declspec_dll int socketTcpReadAll(FD_t sockfd, void* buf, unsigned int nbytes);
+__declspec_dll ssize_t socketReadv(FD_t sockfd, Iobuf_t iov[], unsigned int iovcnt, int flags, struct sockaddr* from, socklen_t* p_slen);
+__declspec_dll ssize_t socketWritev(FD_t sockfd, const Iobuf_t iov[], unsigned int iovcnt, int flags, const struct sockaddr* to, socklen_t tolen);
+__declspec_dll int socketTcpReadAll(FD_t sockfd, void* buf, int nbytes);
 __declspec_dll int socketTcpWriteAll(FD_t sockfd, const void* buf, int nbytes);
 __declspec_dll BOOL socketTcpNoDelay(FD_t sockfd, int on);
 __declspec_dll int socketTcpCanRecvOOB(FD_t sockfd);
